Answers questions in lib/diagnostic.rb
Provides responses to questions for diagnostic. All rake tests are
passed except for the last one, which looks correct per the prompt,
but the test seems to have some ellipses in it, according to my
terminal.

Comfort: 3
Clarity: 4

I'm still getting used to Ruby syntax, but I've gotten a little
more comfortable in the last day or two. I used the Ruby documentation
and yesterday's training notes for this assignment.